Move gateway source into its own folder
This should make code navigation easier and it also standardizes where we need to look for index.js files
