+++ src/gateway/index.js
@@ -4,7 +4,7 @@
  * correct instrumentation of other imported modules.
  */
 // eslint-disable-next-line import/no-unassigned-import
-import "./ka-shared/start-trace-agent.js";
+import "../ka-shared/start-trace-agent.js";
 
 export type {RenderGatewayOptions} from "./types.js";
